MB-31765 handle big docs without getting stuck
Reviewed-on: http://review.couchbase.org/101045
Tested-by: Yu Sui <ysui68@gmail.com>

+++ base/simple_utils.go
@@ -475,7 +475,7 @@ func HasXattr(dataType uint8) bool {
 }
 
 // flatten an array of byte array into a byte array
-func FlattenBytesList(bytesList [][]byte, size int) []byte {
+func FlattenBytesList(bytesList [][]byte, size int64) []byte {
 	flattenedBytes := make([]byte, 0, size)
 	for _, bytes := range bytesList {
 		flattenedBytes = append(flattenedBytes, bytes...)
